backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
replicate==0.23.1
httpx==0.26.0
boto3==1.34.25
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.2.0
openai==1.10.0
elevenlabs==0.2.27
